Added AI
AI can control player's 2 paddle. It will also recalculate the touch_y point after a powerup is taken. It will reposition itself to the center after hitting the ball only if its position is too high or too low. Removed checks of 'touch_y' point in "on_touch_move" function because it is already checked in "update_pos" function of the paddle