Some fixes
Added new ai ball to recalculate touch_y pos after a powerup is taken, so the paddle will move the first time with the first ai ball (aiBall). Removed checks of ball collision from update() function. Instead added a new scheduled event to check it 30 times in a second. (Not 60 otherwise there are more possibilities for the ball getting stuck in the border). Now ai paddle will reset its position after moving to avoid the ball when a yellow powerup is taken